+++ src/Object/User/UserDTO.php
@@ -156,7 +156,7 @@ class UserDTO extends DataTransferObject implements IUserDTO {
 	 * @var array
 	 * @description usr_prop_ilias_roles_info
 	 */
-	protected $iliasRoles = array();
+	protected $iliasRoles = array(self::USER_DEFAULT_ROLE);
 
 
 	/**